# api-documentation.yaml
...
openapi: 3.0.0
info:
  description: This is the API for the Vehicle Cost Calculator
  version: "1.0.0"
  title: Vehicle Cost Calculator API
...
paths:
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Authenticates a User
      description: |
        Authenticates a User and send him a JWT access and refresh token

      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: User123
                password:
                  type: string
                  example: Test-123
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                    format: jwt
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6IlRvbSIsInJvbGUiOiJtZW1iZXIiLCJpYXQiOjE2MzY3NDYwMTcsImV4cCI6MTYzNjc0NjAyN30.CnCV8jVZ5enrhJ5A0ITPoim0z3dDRD2Y64-uUwvNhBk
                  refreshToken:
                    type: string
                    format: jwt
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6IlRvbSIsInJvbGUiOiJtZW1iZXIiLCJpYXQiOjE2MzY3NDYwMTcsImV4cCI6MTYzNjc0NjAyN30.CnCV8jVZ5enrhJ5A0ITPoim0z3dDRD2Y64-uUwvNhBk
        '400':
          description: Request has not the wantet format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Username and Password doesn't match
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/token:
    post:
      tags:
        - Authentication
      summary: Provides a new refresh token
      description: |
        Authenticates a User and send him a JWT access and refresh token

      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
                  format: jwt
                  description: refresh token of the user
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6IlRvbSIsInJvbGUiOiJtZW1iZXIiLCJpYXQiOjE2MzY3NDYwMTcsImV4cCI6MTYzNjc0NjAyN30.CnCV8jVZ5enrhJ5A0ITPoim0z3dDRD2Y64-uUwvNhBk

      responses:
        '200':
          description: Send new access token
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                    format: jwt
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6IlRvbSIsInJvbGUiOiJtZW1iZXIiLCJpYXQiOjE2MzY3NDYwMTcsImV4cCI6MTYzNjc0NjAyN30.CnCV8jVZ5enrhJ5A0ITPoim0z3dDRD2Y64-uUwvNhBk
        '400':
          description: No token sended
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized, refresh token is invalid or expired
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout the User
      description: |
        Removes the given refresh token from the db

      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
                  format: jwt
                  description: refresh token of the user
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6IlRvbSIsInJvbGUiOiJtZW1iZXIiLCJpYXQiOjE2MzY3NDYwMTcsImV4cCI6MTYzNjc0NjAyN30.CnCV8jVZ5enrhJ5A0ITPoim0z3dDRD2Y64-uUwvNhBk

      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Logout successful
        '400':
          description: No token sended
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized, refresh token is invalid or expired
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /users/{username}/vehicles/{vin}/fuelCosts:
    post:
      tags:
        - Fuel Costs
      summary: Add a new Fuel cost item to the Vehicle
      description: |
        Inserts a new Fuel cost item for the given Vehicle into the database
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username of the User
        - in: path
          name: vin
          schema:
            type: string
          required: true
          description: VIN of the Vehicle
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FuelCostItem'
      responses:
        '201':
          description: Fuel cost item created succesfully
        '400':
          description: Request has not the needed format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the needed permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags:
        - Fuel Costs
      summary: Get the Fuel cost item of the vehicle
      description: |
        Delivers all Fuel cost items of the given vehicle
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username of the User
        - in: path
          name: vin
          schema:
            type: string
          required: true
          description: VIN of the Vehicle
      responses:
        '200':
          description: Fuel cost items of the vehicle
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FuelCostItemComplete'
        '400':
          description: Request has not the needed format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the needed permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not able to find user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{username}/vehicles/{vin}/fuelCosts/{id}:
    get:
      tags:
        - Fuel Costs
      summary: Get a specific Fuel cost item of the vehicle
      description: |
        Delivers a specific Fuel cost item of the given id with the givenvin and username
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username of the User
        - in: path
          name: vin
          required: true
          description: Vin of one specific vehicle
          schema:
            type: string
            example: "1P3EW65F4VV300946"
        - in: path
          name: id
          schema:
            type: string
            format: uuid
            example: '56b31f28-6bb2-48dc-b93d-2b6ef0377410'
          required: true
          description: ID of the searched Cost Item
      responses:
        '200':
          description: Fuel cost item of the vehicle
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FuelCostItemComplete'
        '400':
          description: Request has not the needed format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the needed permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not able to find user or vin or id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Fuel Costs
      summary: Deletes one Fuel cost item
      description: |
        Deletes one cost item with the given vin
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username of the User
        - in: path
          name: vin
          required: true
          description: Vin of one specific vehicle
          schema:
            type: string
            example: "1P3EW65F4VV300946"
        - in: path
          name: id
          schema:
            type: string
            format: uuid
            example: '56b31f28-6bb2-48dc-b93d-2b6ef0377410'
          required: true
          description: ID of the searche Cost Item
      responses:
        '204':
          description: Successfully deleted
        '400':
          description: The id does't exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the needed permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not able to find user or vin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Fuel Costs
      summary: Overwrites a Fuel cost item
      description: |
        Overwrites a Fuel cost item
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username of the User
        - in: path
          name: vin
          required: true
          description: Vin of one specific vehicle
          schema:
            type: string
            example: "1P3EW65F4VV300946"
        - in: path
          name: id
          schema:
            type: string
            format: uuid
            example: '56b31f28-6bb2-48dc-b93d-2b6ef0377410'
          required: true
          description: ID of the searche Cost Item
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SingleCostItem'
        description: Fuel cost item to overwrite (vin must match vin in path)
      responses:
        '201':
          description: ID overwritten
        '400':
          description: The Id does't exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the needed permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not able to find user or vin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    FuelCostItem:
      type: object
      required:
        - date
        - price
        - volume
      properties:
        km:
          type: integer
          format: int32
          example: 90600
          description: mileage in km
        price:
          type: number
          description: price in €
          format: float
          example: 80.4
        date:
          type: integer
          format: date
          example: 122234234
          description: gives the date of the cost item in ms
        volume:
          type: number
          format: float
          description: fuel volume in l
          example: 42.5
        consumption:
          type: number
          format: float
          description: consumption of the cost item
        type:
          type: string
          example: "super"
          description: type of the fuel. diesel | super | e10 | super plus | gaz
    FuelCostItemComplete:
      type: object
      required:
        - id
        - vin
        - username
        - date
        - price
        - volume
      properties:
        id:
          type: string
          format: uuid
          example: "6c84fb90-12c4-11e1-840d-7b25c5ee775a"
        vin:
          type: string
          example: "1P3EW65F4VV300946"
        username:
          type: string
          example: "Mustermann25"
        km:
          type: integer
          format: int32
          example: 90600
          description: mileage in km
        price:
          type: number
          description: price in €
          format: float
          example: 80.4
        date:
          type: integer
          format: date
          example: 122234234
          description: gives the date of the cost item in ms
        volume:
          type: number
          format: float
          description: fuel volume in l
          example: 42.5
        consumption:
          type: number
          format: float
          description: consumption of the cost item
        type:
          type: string
          example: "super"
          description: type of the fuel. diesel | super | e10 | super plus | gaz
    SingleCostItem:
      type: object
      required:
        - date
        - price
        - type
        - name
      properties:
        km:
          type: integer
          format: int32
          example: 90600
          description: mileage in km
        price:
          type: number
          description: price in €
          format: float
          example: 80.4
        date:
          type: integer
          format: date
          example: 122234234
          description: gives the date of the cost item in ms
        type:
          type: string
          description: type of the cost item. repair | procurement | rest
          example: "repair"
        name:
          type: string
          description: name of the cost item
          example: "new weels"
        description:
          type: string
          example: "weels for the winter season"
          description: closer description for the cost item (optional)
...
    Error:
      type: object
      required:
        - errors
      properties:
        errors:
          type: array
          items:
            type: object
            required:
              - msg
              - location
            properties:
              value:
                type: string
                example: TestValue
              msg:
                type: string
                example: Invalid Value
              param:
                type: string
                example: Password
              location:
                type: string
                example: body